Completed day 20 part 1

+++ 2022/day-20/src/main/kotlin/com/closeratio/aoc2022/day20/Day20Runner.kt
@@ -10,7 +10,7 @@ class Day20Runner(
 
     override fun getDay(): Int = 20
 
-    override fun part1Function(): () -> Int = {
+    override fun part1Function(): () -> Long = {
         parser
             .parse("/2022_day_20_input.txt")
             .mixAndSum()
